separating methods into TwitchHomeClient
