+++ app/src/main/java/com/example/clicker/network/domain/TwitchRepo.kt
@@ -40,5 +40,4 @@ interface TwitchRepo {
         userId: String
     ):Flow<NetworkAuthResponse<GetModChannels>>
 
-    fun logout(clientId: String, token: String): Flow<Response<String>>
 }
