+++ app/src/test/java/com/example/clicker/ExampleUnitTest.kt
@@ -3,7 +3,6 @@ package com.example.clicker
 import com.example.clicker.network.clients.TwitchClient
 import com.example.clicker.network.domain.TwitchRepo
 import com.example.clicker.network.models.twitchRepo.FollowedLiveStreams
-import com.example.clicker.network.repository.TwitchRepoImpl
 import com.example.clicker.util.Response
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.runTest
