+++ app/src/test/java/com/example/clicker/network/repository/TwitchRepoImplTest.kt
@@ -1,9 +1,10 @@
 package com.example.clicker.network.repository
 
 import android.util.Log
+import com.example.clicker.network.clients.TwitchAuthenticationClient
 import com.example.clicker.network.clients.TwitchClient
 import com.example.clicker.network.domain.TwitchRepo
-import com.example.clicker.network.repository.util.TwitchStreamClientBuilder
+
 import com.example.clicker.presentation.stream.util.Scanner
 import com.example.clicker.presentation.stream.util.Token
 import com.example.clicker.presentation.stream.util.TokenType
@@ -15,12 +16,21 @@ import org.junit.After
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
 
 class TwitchRepoImplTest {
 
     private lateinit var underTest: TwitchRepo
     private lateinit var twitchClient: TwitchClient
     private lateinit var mockWebServer: MockWebServer
+
+    /**WHAT TO TEST*/
+    //1) success with interceptors
+    //2) network interceptor throws errors
+    //3) 401 interceptor throws error
+    //4) empty response body
+    //5) 500 response error
 
 //
     @Before
@@ -36,38 +46,21 @@ class TwitchRepoImplTest {
         mockWebServer.shutdown()
     }
 
-    @Test
-    fun getAllFollowedStreamsNoNetworkResponse() = runTest {
-//        twitchClient = TwitchStreamClientBuilder
-//            .addFailingNetworkInterceptor()
-//            .buildClientWithURL(mockWebServer.url("/").toString()
-//            )
-//        underTest = TwitchRepoImpl(twitchClient)
-//        val expectedResponse = Response.Failure(Exception("Network error, please try again later"))
-//
-//
-//        /**WHEN*/
-//        val actualResponse = underTest
-//            .getFollowedLiveStreams("","","")
-//            .last()
-//
-//
-//        /**THEN*/
-//        Assert.assertEquals(expectedResponse.toString(), actualResponse.toString())
-    }
 
     @Test
-    fun TestingTheLexicalAnanysis(){
-        val scanner = Scanner("/ban @ turkeymeant44 ")
-        scanner.scanTokens()
-        val tokenList =scanner.tokenList
-//        tokenCommand(tokenList)
-
-
-
-        Assert.assertEquals(tokenList.size,4)
-
+    fun `when validateToken() returns a successful response BETTER version`()= runTest{
+        // make the retrofit client
+        val retrofitClient = Retrofit.Builder()
+            .baseUrl(mockWebServer.url("/").toString())
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(TwitchClient::class.java)
+        underTest = TwitchRepoImpl(retrofitClient)
     }
+
+
+
+
